Voici une grosse PR pour l'arrivée de Swan 3.0.0 ! Au programme, un changement d'architecture de projet engendré par le choix d'un changement de framework (Sapphire) par @noftaly, l'ajout de slash commands par @Romitou, d'intéractions super cool par @Romitou et de quelques autres nouveautés par un peu tout le monde :).

- Nouveau message de bienvenue par @SharkZekrom via #174
- ✨ Switch to Sapphire and Discord.js v13 par @noftaly via #175
- ✨ Switch standalone commands to interactions commands par @Romitou via #177
- ✨ Introduce suggestions integration par @Romitou via #178

+++ src/commands/fun/latex.ts
@@ -1,48 +1,54 @@
-import { Command } from 'discord-akairo';
-import type { MessageReaction, User } from 'discord.js';
-import type { GuildMessage } from '@/app/types';
-import type { LatexCommandArguments } from '@/app/types/CommandArguments';
+import type { ChatInputCommand } from '@sapphire/framework';
+import type {
+ ApplicationCommandOptionData, CommandInteraction, MessageReaction, User,
+} from 'discord.js';
+import { Message } from 'discord.js';
+import { ApplicationCommandOptionTypes } from 'discord.js/typings/enums';
+import ApplySwanOptions from '@/app/decorators/swanOptions';
+import SwanCommand from '@/app/structures/commands/SwanCommand';
 import { noop } from '@/app/utils';
 import { latex as config } from '@/conf/commands/fun';
 import messages from '@/conf/messages';
 import settings from '@/conf/settings';
 
-class LatexCommand extends Command {
-  constructor() {
-    super('latex', {
-      aliases: config.settings.aliases,
-      details: config.details,
-      clientPermissions: config.settings.clientPermissions,
-      userPermissions: config.settings.userPermissions,
-      channel: 'guild',
-      args: [{
-        id: 'equation',
-        type: 'string',
-        match: 'content',
-        prompt: {
-          start: config.messages.startPrompt,
-          retry: config.messages.retryPrompt,
-        },
-      }],
-    });
+@ApplySwanOptions(config)
+export default class LatexCommand extends SwanCommand {
+  public static commandOptions: ApplicationCommandOptionData[] = [
+    {
+      type: ApplicationCommandOptionTypes.STRING,
+      name: 'équation',
+      description: 'Équation à mettre en forme en utilisant LaTeX',
+      required: true,
+    },
+  ];
+
+  public override async chatInputRun(
+    interaction: CommandInteraction,
+    _context: ChatInputCommand.RunContext,
+  ): Promise<void> {
+    await this._exec(interaction, interaction.options.getString('équation'));
   }
 
-  public async exec(message: GuildMessage, args: LatexCommandArguments): Promise<void> {
-    const sendMessage = await message.channel.send(settings.apis.latex + encodeURIComponent(args.equation));
+  private async _exec(interaction: CommandInteraction, equation: string): Promise<void> {
+    const sendMessage = await interaction.reply({
+      content: settings.apis.latex + encodeURIComponent(equation),
+      fetchReply: true,
+    });
+    if (!(sendMessage instanceof Message))
+      return;
     await sendMessage.react(settings.emojis.remove).catch(noop);
     const collector = sendMessage
-      .createReactionCollector((reaction: MessageReaction, user: User) => user.id === message.author.id
-        && !user.bot
-        && (reaction.emoji.id ?? reaction.emoji.name) === settings.emojis.remove)
-      .on('collect', async () => {
+      .createReactionCollector({
+        filter: (reaction: MessageReaction, user: User) => user.id === interaction.member.user.id
+          && !user.bot
+          && (reaction.emoji.id ?? reaction.emoji.name) === settings.emojis.remove,
+      }).on('collect', async () => {
         try {
           collector.stop();
           await sendMessage.delete();
         } catch {
-          await message.channel.send(messages.global.oops).catch(noop);
+          await interaction.reply(messages.global.oops).catch(noop);
         }
       });
   }
 }
-
-export default LatexCommand;
